Refactoring: - split tests-api and suite-builder - simplify tests for suite-builder - few new tests for tests-api
