+++ lib/tests-api/suite-builder.js
@@ -1,33 +1,8 @@
 'use strict';
 var _ = require('lodash'),
-    Suite = require('./suite'),
-    State = require('./state');
+    State = require('../state');
 
-function nothing() {}
-
-function notString(arg) {
-    return typeof arg !== 'string';
-}
-
-function argumentsToArray(args) {
-    if (args.length === 1 && Array.isArray(args[0])) {
-        return args[0];
-    } else {
-        return Array.prototype.slice.call(args);
-    }
-}
-
-function isArrayOfStringsAndRegExps(arr) {
-    return _.every(arr, function(item) {
-        return _.isString(item) || _.isRegExp(item);
-    });
-}
-
-function createMatcher(matcher) {
-    return _.isRegExp(matcher)? matcher.test.bind(matcher) : _.isEqual.bind(null, matcher);
-}
-
-function SuiteBuilder(suite) {
+module.exports = function(suite) {
     this.setCaptureElements = function() {
         var selectors = argumentsToArray(arguments);
 
@@ -81,7 +56,7 @@ function SuiteBuilder(suite) {
             opts = null;
         }
 
-        cb = cb || nothing;
+        cb = cb || _.noop;
         opts = opts || {};
 
         if (typeof cb !== 'function') {
@@ -140,42 +115,26 @@ function SuiteBuilder(suite) {
         });
         return this;
     };
+};
+
+function notString(arg) {
+    return typeof arg !== 'string';
 }
 
-module.exports = function(context, suite, browsers) {
-    var suiteId = 1;
-    context.suite = function(name, callback) {
-        if (typeof name !== 'string') {
-            throw new TypeError('First argument of the gemini.suite must be a string');
-        }
+function argumentsToArray(args) {
+    if (args.length === 1 && Array.isArray(args[0])) {
+        return args[0];
+    } else {
+        return Array.prototype.slice.call(args);
+    }
+}
 
-        if (typeof callback !== 'function') {
-            throw new TypeError('Second argument of the gemini.suite must be a function');
-        }
+function isArrayOfStringsAndRegExps(arr) {
+    return _.every(arr, function(item) {
+        return _.isString(item) || _.isRegExp(item);
+    });
+}
 
-        if (suite.hasChildNamed(name)) {
-            throw new Error('Suite ' + name + ' already exists at this level. Choose different name');
-        }
-
-        suite = Suite.create(name, suite);
-        suite.id = suiteId++;
-        if (browsers && suite.parent.isRoot) {
-            suite.browsers = browsers;
-        }
-        callback(new SuiteBuilder(suite));
-        if (suite.hasStates) {
-            if (!suite.url) {
-                throw new Error('URL is required to capture screenshots. ' +
-                    'Call suite.setUrl(<DESIRED PATH>) on "' + suite.name +
-                    '" suite or one of its parents');
-            }
-
-            if (!suite.captureSelectors) {
-                throw new Error('Capture region is required to capture screenshots. ' +
-                    'Call suite.setCaptureElements(<CSS SELECTORS>) on "' + suite.name +
-                    '" suite or one of its parents');
-            }
-        }
-        suite = suite.parent;
-    };
-};
+function createMatcher(matcher) {
+    return _.isRegExp(matcher)? matcher.test.bind(matcher) : _.isEqual.bind(null, matcher);
+}
